+++ classes/class.ilSrUserEnrolmentConfigGUI.php
@@ -4,6 +4,7 @@ require_once __DIR__ . "/../vendor/autoload.php";
 
 use srag\ActiveRecordConfig\SrUserEnrolment\ActiveRecordConfigGUI;
 use srag\Plugins\SrUserEnrolment\Config\ConfigFormGUI;
+use srag\Plugins\SrUserEnrolment\ExcelImport\ExcelImportGUI;
 use srag\Plugins\SrUserEnrolment\Utils\SrUserEnrolmentTrait;
 
 /**
@@ -22,3 +23,5 @@ class ilSrUserEnrolmentConfigGUI extends ActiveRecordConfigGUI {
 		self::TAB_CONFIGURATION => ConfigFormGUI::class
 	];
 }
+
+ExcelImportGUI::CMD_KEY_AUTOCOMPLETE; // TODO: Fix composer autoload (ilCtrl is lowercase and composer not map lowercase :()