+++ src/Log/Log.php
@@ -22,8 +22,10 @@ class Log extends ActiveRecord {
 	use SrUserEnrolmentTrait;
 	const TABLE_NAME = "srusrenr_log";
 	const PLUGIN_CLASS_NAME = ilSrUserEnrolmentPlugin::class;
-	const STATUS_ADD = 100;
-	const STATUS_ERROR = 200;
+	const STATUS_ENROLLED = 100;
+	const STATUS_NOT_ENROLLED = 200;
+	const STATUS_USER_CREATED = 300;
+	const STATUS_USER_UPDATED = 400;
 
 
 	/**
@@ -48,8 +50,10 @@ class Log extends ActiveRecord {
 	 * @var array
 	 */
 	public static $statuss = [
-		self::STATUS_ADD => self::STATUS_ADD,
-		self::STATUS_ERROR => self::STATUS_ERROR
+		self::STATUS_USER_CREATED => self::STATUS_USER_CREATED,
+		self::STATUS_USER_UPDATED => self::STATUS_USER_UPDATED,
+		self::STATUS_ENROLLED => self::STATUS_ENROLLED,
+		self::STATUS_NOT_ENROLLED => self::STATUS_NOT_ENROLLED
 	];
 	/**
 	 * @var int