+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## [2.0.0]
+- Import and enrol users with excel file
+- Reset password from course member list
+- Update description
+
 ## [1.0.0]
 - First version